+++ i18n/get-preffered-language.ts
@@ -1,0 +1,16 @@
+import { NextRequest } from "next/server";
+import { DEFAULT_LANGUAGE, SUPPORTED_LANGUAGES } from "./supported-languages";
+
+export function getPreferredLanguage(request: NextRequest): string {
+  const acceptLanguage = request.headers.get("accept-language");
+  if (!acceptLanguage) return DEFAULT_LANGUAGE;
+
+  const langs = acceptLanguage.split(",").map((lang) => lang.split(";")[0]);
+  for (const lang of langs) {
+    const shortLang = lang.slice(0, 2).toLowerCase();
+    if (SUPPORTED_LANGUAGES.includes(shortLang)) {
+      return shortLang;
+    }
+  }
+  return DEFAULT_LANGUAGE;
+}
