+++ i18n/supported-languages.ts
@@ -1,3 +1,3 @@
-export const SUPPORTED_LANGUAGES = ["en", "ko"] as const;
+export const SUPPORTED_LANGUAGES = ["en", "ko"];
 export type SupportedLanguage = (typeof SUPPORTED_LANGUAGES)[number];
 export const DEFAULT_LANGUAGE = "en";
